+++ test/integration/test_split_fasta.py
@@ -110,3 +110,25 @@ class TestSplitDatabase(TestCaseIntegration):
             'test_circRNA.fasta', 'test_Remaining.fasta', 'test_circRNA.fasta',
             'test_Noncoding.fasta'}
         self.assertEqual(files, expected)
+
+    def test_split_fasta_case4(self):
+        """ test splitFasta case 4 with additional split """
+        args = self.create_base_args()
+        args.gvf = [
+            self.data_dir/'vep/vep_gSNP.gvf',
+            self.data_dir/'vep/vep_gINDEL.gvf',
+            self.data_dir/'reditools/reditools.gvf',
+            self.data_dir/'fusion/star_fusion.gvf',
+            self.data_dir/'circRNA/circ_rna.gvf'
+        ]
+        args.variant_peptides = self.data_dir/'peptides/variant.fasta'
+        args.noncoding_peptides = self.data_dir/'peptides/noncoding.fasta'
+        args.annotation_gtf = self.data_dir/'annotation.gtf'
+        args.proteome_fasta = self.data_dir/'translate.fasta'
+        args.additional_split = ['Noncoding-gSNP']
+        cli.split_fasta(args)
+        files = {str(file.name) for file in self.work_dir.glob('*')}
+        expected = {'test_gINDEL.fasta','test_gSNP.fasta',
+            'test_RNAEditingSite.fasta', 'test_circRNA.fasta',
+            'test_Remaining.fasta', 'test_circRNA.fasta', 'test_Noncoding.fasta'}
+        self.assertEqual(files, expected)